% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{HDZACP-package}
\alias{HDZACP}
\alias{HDZACP-package}
\title{HDZACP: high-dimensional distributional change points}
\description{
HDZACP implements the ZAS (sum), ZAM (maximum), and ZAC (adaptive Cauchy)
rank-scan procedures developed in \emph{Nonparametric Change-Point Detection and
Inference for High-Dimensional Distributions}. The package provides
permutation-calibrated single- and multiple-change procedures, optional
asymptotic calibration, comparison-method wrappers, and complete
reproduction entry points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flnankai/HDZACP}
  \item Report bugs at \url{https://github.com/flnankai/HDZACP/issues}
}

}
\author{
\strong{Maintainer}: Long Feng \email{flnankai@nankai.edu.cn}

Authors:
\itemize{
  \item Ping Zhao
}

}
\keyword{internal}
